Stylistic overhaul and a bug fix
The function was completely revised by cutting down the number
of temporary fields of the Data and Mdata structures, allocating
the all_vars cell array, and using the names cell array only
when float_profs is not an empty array.

The bug fix concerns float and variable selections that resulted
in non-contiguous elements of the params_keep array to be 1. An
out-of-bounds error would occur in this case.