New function check_variables, called from show_sections and show_profiles
The new function will check if the specified variable(s) is/are
defined in Settings.avail_vars. Only defined variables are
returned. As a side effect, if the input was a string, it will
be converted to a cell array, so that after the call to this
function "variables" is always a cell array.