Bug fix: Select float_profs if not all specified float files could be loaded
If not all specified Sprof files could be loaded (either because
they do not exist or because there was a problem with the download)
and if float profiles were specified, only those elements of the
float_profs cell array that correspond to "good" float IDs
have to be selected. This bug fix addresses this issue.