Added return value and optional arguments to list_sensors call
Previously, the function would only display the full list of
sensors used across the entire float array. Now, WMO IDs of
selected floats can be supplied as an argument to show the
sensors that are present on all or some (depending on the
setting of the second argument) of the profiles of the
selected floats.
Also, a cell array with the short names of the sensors found
(regardless of the options used) is returned.