Two bug fixes (varying numbers of sensors across profiles, multiple sensors with mode)
The first fix affects cases where a float has different numbers of
profiles available across profiles. Previously, the first profile
was always used to determine the names of the sensors. Now a profile
with the longest parameter string is used instead.

The second fix affects searches with multiple sensors and the 'mode'
parameter. Previously, this would result in an error. Now each
sensor must be in one of the specified modes.

A third change is due to the fact that in some cases the index file
lists different modes than the corresponding Sprof file. The mode
matching will only be done based on the Sprof files for the time
being.