Allow passing of Data struct as first argument
This change for all wrapper plotting routines allows the use of
the Data struct (as returned by a call to load_float_data) as the
first argument in the function call (instead of float_ids).

This is much faster than loading the Data internally during every
execution of these functions.