Implementation of 'raw','no_strict' option
When option 'raw,'n' is selected, adjusted values will be shown
if they are available for all selected floats. If they are not
available for all floats, raw values will be shown for all.
With the 'raw','no_strict' option, floats without adjusted values
for the selected variables will be omitted from the plots.
Includes a new driver script to illustrate these differences.